requirements.txt
Django==5.0.6
django-health-check==3.18.2
django-advanced-password-validation==1.2.0
django-phonenumber-field==7.3.0
django-configurations==2.5.1
django-cors-headers==4.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gunicorn==22.0.0
whitenoise==6.6.0
setuptools>=67.6.1 # not directly required, pinned by Snyk to avoid a vulnerability
coverage==7.5.2
psycopg2-binary==2.9.9
django-cors-headers==4.4.0

